apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
httpx>=0.27.0
aiohttp>=3.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.35
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
redis>=5.1.0
celery>=5.4.0
litellm>=1.50.0
openai>=1.50.0
anthropic>=0.39.0
pgvector>=0.3.0
PyYAML>=6.0.2
loguru>=0.7.2
tenacity>=9.0.0
cryptography>=43.0.0
argon2-cffi>=23.1.0
playwright>=1.49.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
ruptures>=1.1.9
python-dotenv>=1.0.1
websockets>=13.0
sse-starlette>=2.1.0
streamlit>=1.40.0
plotly>=5.24.0
networkx>=3.3
